Make XP a u32 rather than f32

Closes #226

See merge request veloren/veloren!465

+++ server/src/lib.rs
@@ -315,8 +315,9 @@ impl Server {
                                     if let Some(attacker_stats) = stats.get_mut(attacker) {
                                         // TODO: Discuss whether we should give EXP by Player Killing or not.
                                         attacker_stats.exp.change_by(
-                                            entity_stats.health.maximum() as f64 / 10.0
-                                                + entity_stats.level.level() as f64 * 10.0,
+                                            (entity_stats.health.maximum() as f64 / 10.0
+                                                + entity_stats.level.level() as f64 * 10.0)
+                                                as i64,
                                         );
                                     }
                                 });
